Move some classes around for easier modularization
