+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/LastPublishedLoan.java
@@ -20,7 +20,7 @@ import java.time.OffsetDateTime;
 import java.util.Objects;
 import javax.xml.bind.annotation.XmlElement;
 
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 
 public class LastPublishedLoan extends BaseEntity {
