+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/RawInvestment.java
@@ -28,7 +28,7 @@ import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.InsuranceStatus;
 import com.github.robozonky.api.remote.enums.PaymentStatus;
 import com.github.robozonky.api.remote.enums.Rating;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 
 /**
  * It is not recommended to use this class directly as Zonky will return various null references for fields at various
