+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/Statistics.java
@@ -23,7 +23,7 @@ import java.util.List;
 import javax.xml.bind.annotation.XmlElement;
 
 import com.github.robozonky.api.Ratio;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import io.vavr.Lazy;
 
 public class Statistics extends BaseEntity {
