+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/sanitized/MutableInvestmentImpl.java
@@ -34,8 +34,8 @@ import com.github.robozonky.api.remote.enums.InvestmentStatus;
 import com.github.robozonky.api.remote.enums.PaymentStatus;
 import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.internal.api.Defaults;
-import com.github.robozonky.internal.util.DateUtil;
-import com.github.robozonky.internal.util.RandomUtil;
+import com.github.robozonky.internal.test.DateUtil;
+import com.github.robozonky.internal.test.RandomUtil;
 import com.github.robozonky.internal.util.ToStringBuilder;
 import io.vavr.Lazy;
 import org.apache.logging.log4j.LogManager;
