+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/PrimaryMarketplaceAccessor.java
@@ -27,7 +27,7 @@ import com.github.robozonky.api.strategies.LoanDescriptor;
 import com.github.robozonky.common.remote.Select;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.common.tenant.Tenant;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
