+++ robozonky-app/src/main/java/com/github/robozonky/app/events/impl/AbstractEventImpl.java
@@ -21,7 +21,7 @@ import java.util.stream.Stream;
 
 import com.github.robozonky.api.notifications.Event;
 import com.github.robozonky.common.async.Reloadable;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import com.github.robozonky.internal.util.ToStringBuilder;
 
 /**
