+++ robozonky-app/src/main/java/com/github/robozonky/app/events/impl/LazyEventImpl.java
@@ -21,7 +21,7 @@ import java.util.function.Supplier;
 
 import com.github.robozonky.api.notifications.Event;
 import com.github.robozonky.common.tenant.LazyEvent;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import io.vavr.Lazy;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
