+++ robozonky-app/src/main/java/com/github/robozonky/app/tenant/LoanCache.java
@@ -26,7 +26,7 @@ import java.util.concurrent.ScheduledFuture;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.common.async.Tasks;
 import com.github.robozonky.common.tenant.Tenant;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import io.vavr.Tuple;
 import io.vavr.Tuple2;
 import org.apache.logging.log4j.LogManager;
