+++ robozonky-app/src/main/java/com/github/robozonky/app/transactions/IncomeProcessor.java
@@ -29,7 +29,7 @@ import com.github.robozonky.common.jobs.TenantPayload;
 import com.github.robozonky.common.remote.Select;
 import com.github.robozonky.common.state.InstanceState;
 import com.github.robozonky.common.tenant.Tenant;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
