+++ robozonky-app/src/test/java/com/github/robozonky/app/tenant/ZonkyApiTokenSupplierTest.java
@@ -31,7 +31,7 @@ import com.github.robozonky.common.remote.OAuth;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.common.secrets.SecretProvider;
 import com.github.robozonky.internal.api.Defaults;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import org.junit.jupiter.api.Test;
 import org.mockito.stubbing.Answer;
