+++ robozonky-notifications/src/main/java/com/github/robozonky/notifications/listeners/DelinquencyTracker.java
@@ -19,7 +19,7 @@ package com.github.robozonky.notifications.listeners;
 import com.github.robozonky.api.SessionInfo;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.common.state.TenantState;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.internal.test.DateUtil;
 import com.github.robozonky.notifications.Target;
 
 class DelinquencyTracker {
