+++ robozonky-strategy-natural/src/test/java/com/github/robozonky/strategy/natural/DefaultValuesTest.java
@@ -22,8 +22,8 @@ import java.time.Instant;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.strategies.LoanDescriptor;
 import com.github.robozonky.internal.api.Defaults;
-import com.github.robozonky.internal.util.AbstractMinimalRoboZonkyTest;
-import com.github.robozonky.internal.util.DateUtil;
+import com.github.robozonky.test.AbstractMinimalRoboZonkyTest;
+import com.github.robozonky.internal.test.DateUtil;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
