+++ robozonky-test/src/main/java/com/github/robozonky/test/AbstractRoboZonkyTest.java
@@ -38,7 +38,6 @@ import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.common.state.TenantState;
 import com.github.robozonky.common.tenant.RemotePortfolio;
 import com.github.robozonky.common.tenant.Tenant;
-import com.github.robozonky.internal.util.AbstractMinimalRoboZonkyTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.mockito.stubbing.Answer;
